Move HistoryObjectTemplate into ContainerConfigurationTemplate
